feat: refactor LanguageSwitcher to use new Flags component and remove UKFlag component
